requirements.txt
streamlit
numpy
opencv-python
Pillow
rembg
onnxruntime
tqdm
